Enable dawn_use_built_dxc on Dawn Win x64 bots
This CL enables building DXC on Dawn bots and also enables it
by default on dawn_end2end_tests where available.

To maintain coverage of the FXC path, dawn_end2end_tests is run
with --disable-toggles=use_dxc on these bots as well.

The WebGPU CTS is run with --disable-dawn-features=use_dxc to
keep it on the FXC path. A future change will run the CTS using
DXC.

Note: DXC is not built on Windows x86 because the population of
clients on x86 with SM6.0+ (required for DXIL) is tiny, and
currently blocklisted for WebGPU.

Bug: 1462339
Change-Id: I139a8ad9d019d0dda25de8cc631d050560aeb0c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4663926
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172947}
NOKEYCHECK=True
GitOrigin-RevId: 24454af2a31c5ed96b711f7beac3d0d04a84639b