Update memory metrics immediately on loaded idle
Currently the first metrics request happens after 2 minutes. We
are rolling out a feature to show memory usage in hovercards and it
is confusing to have the memory missing for the first 2 minutes of
a page load.

This change will be behind a FeatureParam associated with the memory
usage in hovercards experiment. On kLoadedIdle we will
request metrics. This might still result in a slight delay before we
have metrics but the data will be higher quality than if we request it
immediately.

Cherry-picked from https://crrev.com/c/4656209.
Original author: Alison Gale <agale@chromium.org>

Bug: 1459009
Change-Id: I0dc85499d039c5498df8f704600688c0e7b0ef95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4701772
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1173549}
NOKEYCHECK=True
GitOrigin-RevId: cae8325b23189c7851e248a399280bd62e6958aa