Fenced frames: Remove unused feature from field trial config.
A previous CL enabled "FencedFramesAPIDefaultMode" for PrivacySandboxV3
trials. This feature does not exist (instead we have
"FencedFramesDefaultMode"), and is only meant to be used in testing, not
out in the wild. This CL removes the feature. There should be no impact
since this was not doing anything to begin with.

Change-Id: I99083fe24831627b2879cf72f6a496dfc20dc3d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4712510
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1174439}
NOKEYCHECK=True
GitOrigin-RevId: 1f05ec664b66682f06f99d85def829dbfce8dfa2