[infra] Parameterize the header for the generated pyl files.
The headers at the top of the generated pyl files are the same except
for which star file should be edited, so this the existing fmt constants
into a single constant that takes the star file as an argument.

This also fixes the header on variants.pyl which mistakenly refers to
mixins.star.

Bug: 1456553
Change-Id: I0a50ef516d58abf7dc7962adabcd2451a3ff1efc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4722358
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1175771}
NOKEYCHECK=True
GitOrigin-RevId: 311ea79f463f2fc45b5ec9f95de668d3917ed758