- Add titouan and pgrace as OWNERS for testing/libfuzzer. - Annotate remaining OWNERS as reviewers of last resort.
When adding new OWNERS I typically talk to folks first and have them be reviewers for the CL to avoid surprises. In this case, @titouan is on leave and I can't ask directly. So I did the next best thing and asked his manager. :) I'm going ahead and proposing the CL, but we can back out the change if it causes anyone troubles.

Also note: pgrace is not yet a committer, but that's the intent so this is simply an optimization of steps on that path.

Change-Id: I19b989d07a89824a404dfd04fcf32a986248fd3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778314
Commit-Queue: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183351}
NOKEYCHECK=True
GitOrigin-RevId: cce0578c458d08131fc166dea0f9c315d972a535