Disable BookmarkBarViewTest2 on Linux ozone/wayland
It tests behaviour that doesn't work on Wayland when not using
xdg_popup.grab. We currently only use xdg_popup.grab when a CLI flag is
passed, i.e. not by default, so we should disable the test.

Note that the test currently passes for reasons that I don't fully
understand, but it will fail as soon as we switch from weston-test to
ui_controls (which is the expected behaviour).

Bug: 1474372
Change-Id: If8f52046f831ab9dfc14ddd14ddb8321fd28f73c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4798588
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Reviewed-by: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1185833}
NOKEYCHECK=True
GitOrigin-RevId: be530b635c29239aac0cf0d6cb1cdafebdac362f