variations: Update `GwpAsanAndroid`
We plan to ramp up GWP-ASan to nonzero sampling. As a prelude to landing
the internal variations config, we update the field trial config to
reflect the "favorite choice" new experimental arm (there are several
new arms).

Quoting from `testing/variations/README.md`:

"the current best-practice is to identify which experiment group is the
most likely candidate for ultimate launch, and to test that
configuration."

Bug: b/287426302
Change-Id: I5e209d40bf2b2d7389d54e97561d3e545ff5e41c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4835198
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192270}
NOKEYCHECK=True
GitOrigin-RevId: db5fcd1d8dcd0802c729fd57ee22c68de2a06922