[omnibox] Delete fieldtrial_testing_config entry for UI Simplification.
Given that we've already launched the 40px row-height variant, this CL
goes ahead and deletes the relevant field trial testing config entry.
This change also ensures that local dev builds of Chromium no longer
see a treatment (the first entry in the deleted config) which is not
going to be launched in the foreseeable future.

Bug: N/A
Change-Id: I03f61e71b7e60e1bc951ba527baacf9f34c6fcf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4832231
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192778}
NOKEYCHECK=True
GitOrigin-RevId: 82cb40c362092a1ff5e5f32c4bd2bb4c742c84d3