Make the AsyncCacheLock feature permanent
The 1% experiment in stable was a success with no negative impacts or
issues.

The actual code path change is in http_cache_transaction.cc (and is 1
change). The rest of the changes are removing the feature flag and
removing the paramaterization of the cache tests.

Bug: 1446082
Change-Id: Ifc651774294a75b6ad3eaf0ef6772325cb63822f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4847127
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193571}
NOKEYCHECK=True
GitOrigin-RevId: f816939b9e7e0bcf2c44a49763de344c870ad2ff