Generate a name field into all test specs.
Currently, the generation of a name field depends on the test type,
whether or not the test had a variant applied or whether the actual test
to be run has been overridden. This requires knowledge of the runtime
behavior to figure out the name of a test when looking at the generated
output.

This CL changes it so that all tests have a name field generated. With
all tests having a name field generated, much of the code can be
simplified because it can assume that the name field is set and just use
that. In particular, there is no need for any test-type-specific
sorting. The name field will have the variant identifier appended when a
test is expanded with a variant, so code that is providing a default
value for the test/isolate will still use the test name set in
test_suites.pyl.

Bug: 1475835
Change-Id: Iada7adfdf94e6747d7dec1a765d4e15fe3cfbe7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4844437
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194259}
NOKEYCHECK=True
GitOrigin-RevId: ffe83c4f20a63f3dce71e743acc578bc7fa7e4e3