Modify ThreadPoolCap feature to apply fixed size
Initial version of the feature restricted certain number of threads
which depended on the number of cores on the device. However, field
data showed that it was better to make a fixed value instead of a
floating value that had less value for higher core architectures.

Bug: b/276797051
Change-Id: I7f341bbe6a6d8eb5e8a77df45ab6c03069abd939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4843951
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Edgar Arriaga García <edgararriaga@google.com>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195690}
NOKEYCHECK=True
GitOrigin-RevId: 453bcb15851b4c0b3309a9758b6de4ccd93f44d6