Treat an empty variants list the same as absent variants list.
Currently, if a matrix compound suite does not specify a variants list
for a basic suite, then non-variant versions of the test get added to
the matrix compound suite. In contrast, a basic suite with an empty
variants list will not have any tests added, which makes the inclusion
of the basic suite in the matrix compound suite pointless. This CL
changes the behavior so that an empty variants list will result in
non-variant versions of the tests in the basic suite being added to the
matrix compound suite. There was a single instance of an empty variants
list, which has been removed to preserve the existing generated json
files.

Change-Id: Ib1146ece3fb0ca742fc7f9f82726dca16a92995e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4863054
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1196302}
NOKEYCHECK=True
GitOrigin-RevId: 60a7b2ae784f39225d6ed62d067125443a691a96