Replace use of isolate_name with test.
The isolate_name field can be set in tests that will be put in the
isolated_scripts test type. The meaning of the field is the same as for
the test field in the gtest test type and the recipe supports either
test or isolate_name being set. Rather than have different fields that
do the same thing in different test types, this CL standardizes on using
test as the name of the field.

Bug: 1475835
Change-Id: I5a25d794bf65af74cc386ce07ebb81f8ce9a9e9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865091
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1196894}
NOKEYCHECK=True
GitOrigin-RevId: dca3d88bd5c0d67cedf0796f007ed6258b1b827d