[ios] use metal backend for graphite on ios-blink-dbg-fyi
Without this change, we revert to Ganesh.

credit: thanks to khushalsagar@ for debugging the text blob pixel
test and noticing that we need to insert the recording before
initiating the async readback (i.e., that the readback isn't part
of the recording itself).

Bug: 1411704,1474022,1482922
Change-Id: Ic168215599f0afbb657b23f951baf1d686d3d760
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4720689
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197223}
NOKEYCHECK=True
GitOrigin-RevId: f16beee29353bc1ae7e41b01cd3a1e046c33978f