[UPMLocalPwd] Enable the migration warning by defalut on non-stable Chrome channels
This CL enables the password migration warning by defalut on all
non-stable Chrome channels and fixes the tests that were failing when
the flag was enabled by defalut.

ChromePasswordManagerClient is passed to DisplaySaveUpdatePasswordPrompt
method of the SaveUpdatePasswordMessageDelegate so that the StubPasswordManagerClient can be passed for testing.
Using the real client was causing the warning to show up for every test.

The prefs version is removed from the fieldtrial testing config and
set to 1 by default because it shouldn't be needed for resetting the
migration warning prefs anymore.

Change-Id: I0c9a52f41b73b4c07b9803262fa5e88997f02a0d
Bug: 1431080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4859225
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1197867}
NOKEYCHECK=True
GitOrigin-RevId: dce9b0012502a4fd821c35854a1762d1c9511f5d