Add field trial testing config for Friendlier Safe Browsing Settings
This is a reland attempt because the original reland
(crrev.com/c/4863865) got reverted due to some of the test cases
failing because our friendlier settings feature wasn't enabled. This CL
adds the EnableFeatures clause to the test cases to ensure the feature
is enabled and the tests don't fail.

Bug: 1465881
Change-Id: I78ee611e834177660c96fdeca3dcf7ecfa0915dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875037
Reviewed-by: Javier Castro <jacastro@chromium.org>
Commit-Queue: Awad Osman <awado@google.com>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198378}
NOKEYCHECK=True
GitOrigin-RevId: d55634f12a4cb9fe32dbdf164d8534aa7803a964