[Switch Access] Don't show setup guide in tests, reducing flakes
The switch access tests in Lacros can flake when the setup guide
is shown and steals focus. This doesn't happen in Ash which usually
completes the whole test before the setup guide is visible. However,
in Lacros, which is slower to launch, the setup guide was stealing
focus in the middle of the test.

The code to keep the setup guide from showing is copied from
TestControllerAsh::SetAssistiveTechnologyEnabled which does the
same work.

AX-Relnotes: N/A
Bug: b/275342946,b/301253962,b/301475127
Change-Id: Ic02ef4c74239084d50811a8b17817bf9c630f751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4886843
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200434}
NOKEYCHECK=True
GitOrigin-RevId: 4914d5be3efcbe0635f06bc43134bd2944c42703