Enforce browser version in GPU tests
Adds the ability for GPU Telemetry tests to enforce that the version of
the browser in use matches the Chromium version for the current
revision. This will catch issues such as the browser not being deployed
correctly on ChromeOS devices.

This is off by default, but enabled for all automated testing.

Bug: 1491879
Change-Id: Idd5b3bc56fbc7aeeea9ea7e0d8f389a69bfa3dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4933049
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1211266}
NOKEYCHECK=True
GitOrigin-RevId: 997e48062611e0d9cd70c8fe423f7883f9f8493f