MiraclePtr: Adjust refcount sizes
On most platforms, we're comfortable with the baked-in default refcount
size. This CL adjusts the field trial config to accept the same on
non-Android platforms.

Android is exempted because we'd like to try enabling the
`all-processes` BRP selector to see how it behaves with 16B refcounts.

Bug: 1444624
Change-Id: I5e3dfaec0175c8cb7adcd9454d8038fde2f61899
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4904416
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212594}
NOKEYCHECK=True
GitOrigin-RevId: 812f9ec65fcd771bcac7ce42e249a48482e5ba51