[Tab Switcher] Change Android Animation start timeout to a param
The existing timeout of 250 seems too restrictive on emulators. Raise
to 300 and make it a parameter so we can modify it at rollout.

This CL looks a lot bigger than it is due to auto reformatting.

Bug: 1444782
Change-Id: Ic5afa021a45ab0e9a2df4437f408abfba93939aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4975038
Reviewed-by: Xi Han <hanxi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1214867}
NOKEYCHECK=True
GitOrigin-RevId: c2a08a124a665fd2285914410a3f606f0a3a58db