Add new enable_fuzztest option to test.gni
For ClusterFuzz to run the executables that contain FUZZ_TESTs, it needs
to know to pass -fuzz=. This CL generates a .options file for any such
test. This requires the underlying test.gni template unfortunately to
know a little about FuzzTest, so it knows to create the .options file.

Bug: 1494445
Change-Id: I31b0f8784620b7eb5b49522da7134cdfee0d317e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4966725
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Paul Semel <paulsemel@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215057}
NOKEYCHECK=True
GitOrigin-RevId: 62dbea584ecc98c698bfccf4391ccf2536e2aff6