[IntersectionOptimization] Disable fieldtrial testing config
I will make some changes to the feature and it would be risky to
enable it in fieldtrial testing config. Now let it follow the default
status ("experimental" in runtime_enabled_features.json5). It'll be
still enabled for web tests by default, with
virtual/disable-intersection-optimization for test coverage of the
disabled code path. It'll be disabled by default in other tests.

Bug: 1400495
Change-Id: I0a2887657f96e30ddac678aa3a7120622e932a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4985461
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218339}
NOKEYCHECK=True
GitOrigin-RevId: 2a24d430ee4ea5c8e1e88fb5013f790cb2fe90b1