Reland "Enable retry_only_failed_tests for browser_tests"
This is a reland of commit 698feb84bf081b63e2550f1c09c8c1dc13001eb6

This didn't need to be reverted as the recipe is functionally
correct, re-running only the non-exonerated failing tests.
The UI is just confusing with an extra harmless shard that runs
with no tests when a test is exonerated. Filed crbug/1499972
to fix that.

Original change's description:
> Enable retry_only_failed_tests for browser_tests
>
> Enables for linux-rel, linux-wayland-rel, win-rel, linux-lacros-rel.
> Previously rolled back due to bugs but fixed in https://chromium-review.googlesource.com/c/chromium/tools/build/+/4848488
>
> Bug: 1473501
> Change-Id: Ibda1949b20402327d623576da7c97a731c6b6426
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4856429
> Commit-Queue: Stephanie Kim <kimstephanie@google.com>
> Reviewed-by: Garrett Beaty <gbeaty@google.com>
> Cr-Commit-Position: refs/heads/main@{#1194977}

Bug: 1473501
Change-Id: Ibe82b9deac0853156433bfd1c541efbbc5a1a3b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5007538
Commit-Queue: Stephanie Kim <kimstephanie@google.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1221033}
NOKEYCHECK=True
GitOrigin-RevId: 3df2d0c8e6e2308df59d37a18aae1a21a218e7e1