[flake suppressor improvement] 2/N: Add Skip status for slow test in Expectation File.
Added two new fields to test results.
is_slow: used to know if test is slow for this run, this is added
for future slow test detection needs.
typ_expectations: List of the expected test run result.

If Slow is in typ_expectations, the test is already in
Slow TestExpectations file, which need to add the Skip status.

Bug: 1358735
Change-Id: I2264dffbf1ed0f3a062d95e82a8a779c0ae7ddb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5004261
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Jie Sheng <jiesheng@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1221244}
NOKEYCHECK=True
GitOrigin-RevId: 2b23b6bc3fa3905c038f616f76a8c0fcf8d7752d