Prefetch: Enable PrefetchNewLimits by default
Enables the feature by default and removes the
fieldtrial_testing_config for the feature; in preparation roll out to
100% of Stable users. See bug for Stable experiment results.

This also includes a test fix in preloading_decider_unittest.cc;
we remove an explicit call to
PreloadingDecider::GetForCurrentDocument() as that method is called
in ~PrefetchContainer() when ResetPrefetch is called. The test
using it should have been guarded behind PrefetchNewLimits to begin
with (CandidateCanBeReprefetchedAfterDiscard).

Bug: 1445086
Change-Id: Idcafd011686b6a2b2c56208fb18f21bc74e495f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5037590
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226868}
NOKEYCHECK=True
GitOrigin-RevId: 62769b27b011d93580a1ae740ab2e8ced59e566e